requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.2
python-dotenv>=1.0.1
google-generativeai>=0.8.3
pydantic-settings>=2.6.0
genkit>=0.4.0
firebase-admin>=6.5.0
httpx>=0.28.1
firebase-admin>=6.5.0
google-cloud-firestore>=2.21.0